Handler refactoring, step 1: YoutubeHandler and WebsiteHandler now both implement IURLHandler and are invoked as such
