+++ DeathmicChatbot/Handlers/YoutubeHandler.cs
@@ -7,14 +7,14 @@ using Google.YouTube;
 #endregion
 
 
-namespace DeathmicChatbot
+namespace DeathmicChatbot.Handlers
 {
-	internal class YotubeManager : IURLHandler
+	internal class YoutubeHandler : IURLHandler
     {
         private readonly Regex _reg;
         private readonly YouTubeRequest _request;
 
-        public YotubeManager()
+        public YoutubeHandler()
         {
             var settings = new YouTubeRequestSettings("Youtube Title Bot",
                                                       "AI39si6DFGChi5M0rnrX6p5dasT6STlFELYpJdbxdVXR3L1-Cj5RzNUU2nsm2LPmshlVGHuYmeaZ30zGJgqdhSSNoWQgJmEEDA");
